feat: Migrate almost all CSS to BEM style `elm-css`
With the exception of `..:before { content: " " }` we can model everything we
need in `elm-css`. Exposing BEM modifiers via the API only seems to make sense
with something like `button` and `pill` where we potentially face a lot of
combinations - doing this with `panelMenu` feels like overkill right now, we
might reconsider in case we have more than the `footer` and `header` variants.

It's a bit tedious to have all that boilerplate in `App.UI` to map the BEM
custom types to a name, put the actual CSS for these into the stylesheet
and expose one or more functions and zero or more modifier types per element
but the usage side of the API makes a very clean expression.